+++ opensquirrel/ir/default_gates/single_qubit_gates.py
@@ -4,7 +4,7 @@ from typing import SupportsFloat
 import numpy as np
 
 from opensquirrel.ir import Axis, AxisLike, QubitLike
-from opensquirrel.ir.semantics import BsrAngleParam, BsrFullParams, BsrNoParams
+from opensquirrel.ir.semantics import BsrAngleParam, BsrFullParams, BsrNoParams, BsrUnitaryParams
 
 
 class Rn(BsrFullParams):
@@ -111,7 +111,7 @@ class TDagger(BsrNoParams):
         BsrNoParams.__init__(self, qubit=qubit, axis=(0, 0, 1), angle=-pi / 4, phase=-pi / 8, name="Tdag")
 
 
-class U(BsrFullParams):
+class U(BsrUnitaryParams):
     def __init__(
         self,
         qubit: QubitLike,
@@ -119,11 +119,4 @@ class U(BsrFullParams):
         phi: SupportsFloat,
         lmbda: SupportsFloat,
     ) -> None:
-        from opensquirrel.passes.merger.general_merger import compose_bloch_sphere_rotations  # lazy import
-
-        a = Rn(qubit, 0, 0, 1, lmbda, phi=0)
-        b = Rn(qubit, 0, 1, 0, theta, phi=0)
-        c = Rn(qubit, 0, 0, 1, phi, phi=(float(phi) + float(lmbda)) / 2)
-        bsr = compose_bloch_sphere_rotations(compose_bloch_sphere_rotations(a, b), c)
-
-        BsrFullParams.__init__(self, qubit=qubit, axis=bsr.axis, angle=bsr.angle, phase=bsr.phase, name="U")
+        BsrUnitaryParams.__init__(self, qubit=qubit, theta=theta, phi=phi, lmbda=lmbda, name="U")